Moving defn & impl for GameState into its own file. Cleanup for both types.rs & game.rs
